third_party/blink/web_tests/external/wpt/html/cross-origin-embedder-policy/reporting-to-endpoint.https-expected.txt

This is a testharness.js-based test.
PASS subresource CORP
PASS navigation CORP
PASS COEP violation on nested frame navigation
PASS Two COEP headers, split inside report-to value
FAIL Shared worker fetch assert_equals: expected "iframe" but got ""
Harness: the test ran to completion.
